flask_app/requirements.txt
Flask==2.2.5
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.29
psycopg2-binary==2.9.7
PyMySQL==1.0.3
kafka-python==2.0.2
requests==2.31.0
python-dotenv==1.0.1
gunicorn==20.1.0
Werkzeug==2.2.3